Правки компонентов, заготовка логики под бэкенд
